feat: Add SmoothTransitionWithDefaultConfigs
Introduce a new wrapper component, SmoothTransitionWithDefaultConfigs,
for the client component SmoothTransition, enabling server rendering
with default configs from configsTransition. Suitable for scenarios
requiring only default configurations. Include a basic unit test for the
new component.

Clean up transition types by merging smoothTransition.types with
transition.types. Rename and relocate smoothTransition.data, and
remove unused types.
