+++ app/_components/ui/transition/smoothTransitionWithDivRef/__test__/smoothTransitionWithDivRef.test.tsx
@@ -1,10 +1,18 @@
 import { render, screen } from '@testing-library/react';
 import { SmoothTransitionWithDivRef } from '..';
 import { ReactNode } from 'react';
+import { configsTransition } from '@/transition/transition.configs';
 
 describe('SmoothTransitionWithDivRef', () => {
   const renderWithSmoothTransitionWithDivRef = (children: ReactNode) =>
-    render(<SmoothTransitionWithDivRef _id={null}>{children}</SmoothTransitionWithDivRef>);
+    render(
+      <SmoothTransitionWithDivRef
+        configs={configsTransition()}
+        _id={null}
+      >
+        {children}
+      </SmoothTransitionWithDivRef>,
+    );
 
   it('should render the child elements', async () => {
     const { container } = renderWithSmoothTransitionWithDivRef(<div>smoothTransition-test</div>);
