+++ app/_components/ui/transition/smoothTransitionWithDivRef/index.tsx
@@ -3,13 +3,9 @@
 import { useAtomValue } from 'jotai';
 import { atomDivRef } from '@/container/container.states';
 import { SmoothTransition } from '../smoothTransition';
-import { PropsSmoothTransitionWithDivRef } from '../smoothTransition/smoothTransition.types';
+import { PropsSmoothTransitionWithDivRef } from '../transition.types';
 
-export const SmoothTransitionWithDivRef = ({
-  children,
-  configs,
-  _id = null,
-}: PropsSmoothTransitionWithDivRef) => {
+export const SmoothTransitionWithDivRef = ({ children, configs, _id = null }: PropsSmoothTransitionWithDivRef) => {
   const divRef = useAtomValue(atomDivRef(_id));
 
   return (
