+++ app/v2/page.tsx
@@ -1,5 +1,17 @@
+import { SectionContent } from '@/section/sectionContent';
+import { SectionHeader } from '@/section/sectionHeader';
+import { SectionHero } from '@/section/sectionHero';
+import { SectionStartToday } from '@/section/sectionStartToday';
+
 const Home = () => {
-  return <div>Home v2</div>;
+  return (
+    <>
+      <SectionHero />
+      <SectionHeader />
+      <SectionContent />
+      <SectionStartToday />
+    </>
+  );
 };
 
 export default Home;
